# Optimize traverse_and_update without accumulator